requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.16.55
botocore==1.19.55
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-import-export==2.5.0
django-storages==1.11.1
et-xmlfile==1.0.1
gevent==21.1.2
greenlet==1.0.0
grequests==0.6.0
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.2.0
